Overhauled entire application.
The original program was fundamentally flawed; plain and simple:
In an effort to avoid using the infamous "global variable", I made the task twice as hard as it needed to be. My first mistake was not initializing the testScore member of the StudentProfile structure as an array in the first place. This lead me to devise a crude solution; create a separate member variable and initialize that as an array to collect the scores...only that's not how structures quite work, I believe.

Anyway, the resultant worked...but would cause memory leaks. Or something of that nature. Ether way, when the compiler has to break your code, that's usually a red flag that something is wrong.

Global CONSTANT variables are the exception to the rule!

Once I initialized the testScore member as an array, everything else seemed to effortlessly fall into place.

Thank you Melvin, wherever the hell you are.